+++ inst/include/ns/ns_RCM.hpp
@@ -9,16 +9,16 @@ Type log2(Type x) {
 }
 
 template<class Type>
-matrix<Type> generate_ALK(vector<Type> length_bin, matrix<Type> len_age, matrix<Type> SD_LAA, 
-                          int n_age, int nlbin, Type bin_width, int y) {
-  matrix<Type> ALK(n_age, length_bin.size());
+matrix<Type> generate_ALK(vector<Type> lbin, matrix<Type> len_age, matrix<Type> SD_LAA, 
+                          int n_age, int nlbin, int y) {
+  matrix<Type> ALK(n_age, nlbin);
   for(int a=0;a<n_age;a++) {
     for(int j=0;j<nlbin;j++) {
       if(j==nlbin-1) {
-        ALK(a,j) = 1 - pnorm(length_bin(j) - 0.5 * bin_width, len_age(y,a), SD_LAA(y,a));
+        ALK(a,j) = 1 - pnorm(lbin(j), len_age(y,a), SD_LAA(y,a));
       } else {
-        ALK(a,j) = pnorm(length_bin(j) + 0.5 * bin_width, len_age(y,a), SD_LAA(y,a));
-        if(j>0) ALK(a,j) -= pnorm(length_bin(j) - 0.5 * bin_width, len_age(y,a), SD_LAA(y,a));
+        ALK(a,j) = pnorm(lbin(j+1), len_age(y,a), SD_LAA(y,a));
+        if(j>0) ALK(a,j) -= pnorm(lbin(j), len_age(y,a), SD_LAA(y,a));
       }
     }
   }